internal: Introduce FlowOp marker trait for flow operators
Move flow-related AST nodes (StageDef, FlowRoute, FlowFork, FlowMerge,
FlowWait, FlowActivate, FlowJump, FlowEnd) from relation.scala to a
dedicated flow.scala file.

Introduce FlowOp as a marker trait that extends Relation. This provides:
- Semantic clarity: Flow operators can be identified via pattern matching
- Compatibility: Works with existing parser pipe chain infrastructure
- Code organization: Flow operators are now in a dedicated file
